requirements.txt
attrs==24.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
lxml==5.2.2
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pytest==8.3.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.2.2
zeep==4.2.1
